refactory: rename button menu
