+++ libs/feature/src/lib/components/menu/index.ts
@@ -1,3 +1,3 @@
 export * from './base-menu';
-export * from './button-file-menu';
+export * from './button-menu';
 export * from './right-click-menu';
